chore: keyboardshortcuts rewrite
