+++ apps/meteor/client/views/room/contextualBar/KeyboardShortcuts/KeyboardShortcutSection.tsx
@@ -1,0 +1,19 @@
+import { Box, Divider } from '@rocket.chat/fuselage';
+import React, { ReactElement } from 'react';
+
+type KeyboardShortcutSectionProps = {
+	title: string;
+	command: string;
+};
+
+const KeyboardShortcutSection = ({ title, command }: KeyboardShortcutSectionProps): ReactElement => (
+	<Box is='section' mb='x16'>
+		<Box fontScale='p2m' fontWeight='700'>
+			{title}
+		</Box>
+		<Divider />
+		<Box fontScale='p2'>{command}</Box>
+	</Box>
+);
+
+export default KeyboardShortcutSection;
