--- apps/meteor/client/views/room/contextualBar/KeyboardShortcuts/KeyboardShortcutsWithClose.js
@@ -1,11 +1,0 @@
-import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
-import React, { memo } from 'react';
-
-import KeyboardShortcuts from './KeyboardShortcuts';
-
-const KeyboardShortcutsWithClose = ({ tabBar }) => {
-	const handleClose = useMutableCallback(() => tabBar && tabBar.close());
-	return <KeyboardShortcuts handleClose={handleClose} />;
-};
-
-export default memo(KeyboardShortcutsWithClose);
