+++ apps/meteor/client/views/room/contextualBar/KeyboardShortcuts/index.ts
@@ -1,1 +1,1 @@
-export { default } from './KeyboardShortcutsWithClose';
+export { default } from './KeyboardShortcuts';
